fix(union-eyes): make Django AddField idempotent against platform SQL 0005/0006
Platform SQL from #799 already adds the overlapping organization_members
and documents columns. Record those columns in Django state and apply them
with ADD COLUMN IF NOT EXISTS using the platform types, including
documents.file_size as bigint. Keep real Django AddField for columns that
0005/0006 do not create, including claims.idempotency_hash.

+++ apps/union-eyes/backend/content/models.py
@@ -288,7 +288,9 @@ class Documents(BaseModel):
     mime_type = models.TextField(null=True, blank=True)
     document_type = models.TextField(null=True, blank=True)
     size_bytes = models.BigIntegerField(null=True, blank=True)
-    file_size = models.IntegerField(null=True, blank=True)
+    # bigint matches PLATFORM_SQL 0006 (documents.file_size). IntegerField would
+    # materialize a conflicting integer when Django runs before that migration.
+    file_size = models.BigIntegerField(null=True, blank=True)
     category = models.TextField(null=True, blank=True)
     tags = ArrayField(models.TextField(), null=True, blank=True)
